adding unit tests for determined fired rules
