+++ src/main/java/org/occideas/interview/rest/InterviewRestController.java
@@ -2,24 +2,15 @@ package org.occideas.interview.rest;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.occideas.agent.service.AgentService;
 import org.occideas.base.rest.BaseRestController;
 import org.occideas.entity.Interview;
 import org.occideas.entity.InterviewAnswer;
 import org.occideas.entity.InterviewQuestion;
 import org.occideas.fragment.service.FragmentService;
 import org.occideas.interview.service.InterviewService;
-import org.occideas.interviewanswer.service.InterviewAnswerService;
-import org.occideas.interviewautoassessment.service.InterviewAutoAssessmentService;
-import org.occideas.interviewfiredrules.service.InterviewFiredRulesService;
-import org.occideas.interviewmanualassessment.service.InterviewManualAssessmentService;
 import org.occideas.interviewmodule.service.InterviewModuleService;
 import org.occideas.module.service.ModuleService;
-import org.occideas.modulerule.service.ModuleRuleService;
 import org.occideas.question.service.QuestionService;
-import org.occideas.systemproperty.service.SystemPropertyService;
-import org.occideas.utilities.AssessmentStatusEnum;
-import org.occideas.utilities.StudyAgentUtil;
 import org.occideas.vo.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
@@ -39,40 +30,16 @@ public class InterviewRestController implements BaseRestController<InterviewVO> 
   private InterviewService service;
 
   @Autowired
-  private InterviewAutoAssessmentService autoAssessmentService;
-
-  @Autowired
-  private InterviewManualAssessmentService manualAssessmentService;
-
-  @Autowired
-  private InterviewFiredRulesService firedRulesService;
-
-  @Autowired
-  private InterviewAnswerService interviewAnswerService;
-
-  @Autowired
   private QuestionService questionService;
 
   @Autowired
   private InterviewModuleService modService;
 
   @Autowired
-  private AgentService agentService;
-
-  @Autowired
-  private ModuleRuleService moduleRuleService;
-
-  @Autowired
-  private SystemPropertyService sysPropService;
-
-  @Autowired
   private ModuleService moduleService;
 
   @Autowired
   private FragmentService fragmentService;
-
-  @Autowired
-  private StudyAgentUtil studyAgentUtil;
 
   @GET
   @Path(value = "/getlist")
@@ -274,83 +241,13 @@ public class InterviewRestController implements BaseRestController<InterviewVO> 
   @Path(value = "/updateAutoAssessments")
   @Produces(value = MediaType.APPLICATION_JSON_VALUE)
   public Response updateAutoAssessments() {
-    List<InterviewVO> list = service.listAllInterviewsWithoutAnswers();
     try {
-
-      List<AgentVO> listAgents = agentService.getStudyAgents();
-      int iSize = list.size();
-      int iCount = 0;
-      for (InterviewVO interviewVO : list) {
-        iCount++;
-        log.info("auto assessment for " + interviewVO.getInterviewId() + " triggered " + iCount + " of " + iSize);
-        System.out.println("ID " + interviewVO.getInterviewId() + " triggered " + iCount + " of " + iSize);
-        interviewVO = this.determineFiredRules(interviewVO);
-        // get interviewautoassessment for each interview
-        // can change if we can do a batch retrieve for all interviews
-        List<InterviewAutoAssessmentVO> assessments = autoAssessmentService.findByInterviewId(interviewVO.getInterviewId());
-        ArrayList<RuleVO> autoAssessedRules = new ArrayList<RuleVO>();
-        for (InterviewAutoAssessmentVO assessment : assessments) {
-          assessment.getRule().setDeleted(1);
-          autoAssessedRules.add(assessment.getRule());
-        }
-        List<InterviewFiredRulesVO> interviewFiredRules = firedRulesService.findByInterviewId(interviewVO.getInterviewId());
-        ArrayList<RuleVO> firedRules = new ArrayList<RuleVO>();
-        for (InterviewFiredRulesVO ifiredRule : interviewFiredRules) {
-          RuleVO firedRule = ifiredRule.getRules().get(0);
-          firedRules.add(firedRule);
-        }
-
-        setRuleLevelNoExposure(listAgents, autoAssessedRules, firedRules);
-        evaluateAssessmentStatus(interviewVO);
-
-        interviewVO.setAutoAssessedRules(autoAssessedRules);
-        System.out.println("Status is" + interviewVO.getAssessedStatus());
-        service.update(interviewVO);
-
-      }
+      List<InterviewVO> results = service.autoAssessedRules();
       log.info("Assessment completed for interviewIds");
+      return Response.ok(results).build();
     } catch (Throwable e) {
       log.error("Error Occured on assessment", e.getMessage());
       return Response.status(Status.BAD_REQUEST).type("text/plain").entity(e.getMessage()).build();
-    }
-    return Response.ok(list).build();
-  }
-
-  private void evaluateAssessmentStatus(InterviewVO interviewVO) {
-    if (hasManualAssessedRules(interviewVO)) {
-      interviewVO.setAssessedStatus(AssessmentStatusEnum.MANUALLYASSESSED.getDisplay());
-    } else {
-      if (interviewVO.getAssessedStatus().equalsIgnoreCase(AssessmentStatusEnum.NOTASSESSED.getDisplay())) {
-        interviewVO.setAssessedStatus(AssessmentStatusEnum.AUTOASSESSED.getDisplay());
-      } else {
-        //leave status the same for now
-      }
-
-    }
-  }
-
-  private boolean hasManualAssessedRules(InterviewVO interviewVO) {
-    return interviewVO.getManualAssessedRules() != null &&
-      !interviewVO.getManualAssessedRules().isEmpty();
-  }
-
-  private void setRuleLevelNoExposure(List<AgentVO> listAgents, ArrayList<RuleVO> autoAssessedRules, ArrayList<RuleVO> firedRules) {
-    for (AgentVO agent : listAgents) {
-      RuleVO rule = new RuleVO();
-      rule.setLevel("noExposure");
-      rule.setLevelValue(5);
-      rule.setAgentId(agent.getIdAgent());
-      for (RuleVO ruleFired : firedRules) {
-        if (ruleFired.getAgentId() == agent.getIdAgent()) {
-          if (ruleFired.getLevelValue() <= rule.getLevelValue()) {
-            rule.setLevel(ruleFired.getLevel());
-            rule.setLevelValue(ruleFired.getLevelValue());
-            rule.setAgentId(ruleFired.getAgentId());
-            rule.setLegacyRuleId(ruleFired.getIdRule());
-          }
-        }
-      }
-      autoAssessedRules.add(rule);
     }
   }
 
@@ -377,7 +274,7 @@ public class InterviewRestController implements BaseRestController<InterviewVO> 
     try {
       list = service.findByIdWithRules(id);
       for (InterviewVO interviewVO : list) {
-        interviewVO = this.determineFiredRules(interviewVO);
+        interviewVO = service.determineFiredRules(interviewVO);
       }
     } catch (Throwable e) {
       e.printStackTrace();
@@ -517,87 +414,6 @@ public class InterviewRestController implements BaseRestController<InterviewVO> 
 
   private QuestionVO getNearestQuestion(String parentId, String number) {
     return questionService.determineNextQuestionByCurrentNumber(parentId, number);
-  }
-
-  private InterviewVO determineFiredRules(InterviewVO interview) {
-    ArrayList<RuleVO> firedRules = new ArrayList<RuleVO>();
-    ArrayList<RuleVO> rules = new ArrayList<RuleVO>();
-    // TODO need to move this call every after interview that is finished
-    service.cleanDeletedAnswers(interview.getInterviewId());
-    // get all answers by interviewId
-
-    /**
-     * TODO START refactor to use 1 sql query to get the answerhistory joined with modulerule table
-     */
-    List<InterviewAnswerVO> answerHistory = interviewAnswerService.findByInterviewId(interview.getInterviewId());
-    //h
-    List<ModuleRuleVO> moduleRules = new ArrayList<ModuleRuleVO>();
-    for (InterviewAnswerVO ia : answerHistory) {
-      //for(InterviewAnswerVO ia:iq.getAnswers()){
-      List<ModuleRuleVO> nodeModuleRules = moduleRuleService.findByIdNode(ia.getAnswerId());
-
-      moduleRules.addAll(nodeModuleRules);
-
-      //}
-    }
-    /**
-     * TODO END refactor to use 1 sql query to get the answerhistory joined with modulerule table
-     */
-    for (ModuleRuleVO moduleRule : moduleRules) {
-      if (!rules.contains(moduleRule.getRule())) {
-        rules.add(moduleRule.getRule());
-      }
-    }
-    //get list of answer nodes
-    List<InterviewAnswerVO> allActualAnswers = answerHistory;
-    allActualAnswers = removeDeletedAnswers(allActualAnswers);
-    for (RuleVO rule : rules) {
-      boolean bFired = false;
-      for (PossibleAnswerVO pa : rule.getConditions()) {
-        boolean bPartialFired = false;
-        for (InterviewAnswerVO ia : allActualAnswers) {
-          if (pa.getIdNode() == ia.getAnswerId()) {
-            bPartialFired = true;
-            bFired = true;
-            break;
-          }
-        }
-        if (!bPartialFired) {
-          bFired = false;
-          break;
-        }
-      }
-      if (bFired) {
-        if (!firedRules.contains(rule)) {
-          firedRules.add(rule);
-        }
-      }
-    }
-    interview.setFiredRules(firedRules);
-    NoteVO note = new NoteVO();
-    note.setInterviewId(interview.getInterviewId());
-    note.setText("Ran determineFiredRules");
-    note.setType("System");
-    interview.getNotes().add(note);
-    List<InterviewManualAssessmentVO> assessmentRules = manualAssessmentService.findByInterviewId(interview.getInterviewId());
-    ArrayList<RuleVO> manualAssessedRules = new ArrayList<RuleVO>();
-    for (InterviewManualAssessmentVO assessment : assessmentRules) {
-      manualAssessedRules.add(assessment.getRule());
-    }
-    interview.setManualAssessedRules(manualAssessedRules);
-    //System.out.print("Assessment status before update:"+interview.getAssessedStatus());
-    service.update(interview);
-    return interview;
-  }
-
-  private List<InterviewAnswerVO> removeDeletedAnswers(List<InterviewAnswerVO> allActualAnswers) {
-    ArrayList<InterviewAnswerVO> retValue = new ArrayList<InterviewAnswerVO>();
-    for (InterviewAnswerVO answer : allActualAnswers) {
-      if (answer.getDeleted() == 0) {
-        retValue.add(answer);
-      }
-    }
-    return retValue;
   }
 
   private ArrayList<RuleVO> removeDuplicates(List<RuleVO> rules) {
